+++ R/batch_geocoding.R
@@ -3,11 +3,11 @@
 # Census batch geocoding
 # @param address_pack packaged addresses object
 # Vintage must be defined if return = 'geographies'
-batch_census <- function(address_pack,
+batch_census <- function(unique_addresses,
      return_type = 'locations', timeout = 20, full_results = FALSE, custom_query = list(), api_url = NULL,
      lat = 'lat', long = 'long', verbose = FALSE, ...) {
   
-  if (!'street' %in% names(address_pack$unique) & (!'address' %in% names(address_pack$unique))) {
+  if (!'street' %in% names(unique_addresses) & (!'address' %in% names(unique_addresses))) {
     stop("To use the census geocoder, either 'street' or 'address' must be defined")
   }
   
@@ -20,15 +20,15 @@ batch_census <- function(address_pack,
   
   if (is.null(api_url)) api_url <- get_census_url(return_type, 'addressbatch')
   
-  num_addresses <- nrow(address_pack$unique)
+  num_addresses <- nrow(unique_addresses)
 
   # create input dataframe
   input_df <- tibble::tibble(
     id      = 1:num_addresses,
-    street  = if ('street' %in% names(address_pack$unique)) address_pack$unique$street else address_pack$unique$address,
-    city    = if ('city' %in% names(address_pack$unique)) address_pack$unique$city else NA,
-    state   = if ('state' %in% names(address_pack$unique)) address_pack$unique$state else NA,
-    zip     = if ('postalcode' %in% names(address_pack$unique)) address_pack$unique$postalcode else NA
+    street  = if ('street' %in% names(unique_addresses)) unique_addresses$street else unique_addresses$address,
+    city    = if ('city' %in% names(unique_addresses)) unique_addresses$city else NA,
+    state   = if ('state' %in% names(unique_addresses)) unique_addresses$state else NA,
+    zip     = if ('postalcode' %in% names(unique_addresses)) unique_addresses$postalcode else NA
   )
   
   # Write a Temporary CSV
@@ -51,7 +51,7 @@ batch_census <- function(address_pack,
          'county_fips' = 'character',
          'census_tract' = 'character',
          'census_block' = 'character'),
-       NA)
+          NA)
   
   results <- utils::read.csv(text = raw_content, header = FALSE,
        col.names = return_cols,
@@ -59,7 +59,11 @@ batch_census <- function(address_pack,
        fill = TRUE, stringsAsFactors = FALSE,
        na.strings = '')
   
-  results <- results[order(results[['id']]), ]  # make sure results remain in proper order
+  # convert 'id' to integer since we sort on it
+  results[['id']] <- as.integer(results[['id']])
+  
+  # make sure results remain in proper order
+  results <- results[order(results[['id']]), ]
 
   # split out lat/lng. lapply is used with as.numeric to convert coordinates to numeric
   coord_df <- do.call(rbind, lapply(results$coords, split_coords))
@@ -79,12 +83,12 @@ batch_census <- function(address_pack,
 
 # Batch geocoding with geocodio
 # ... are arguments passed from the geo() function
-batch_geocodio <- function(address_pack, lat = 'lat', long = 'long', timeout = 20, full_results = FALSE, custom_query = list(),
+# https://www.geocod.io/docs/#batch-geocoding
+batch_geocodio <- function(unique_addresses, lat = 'lat', long = 'long', timeout = 20, full_results = FALSE, custom_query = list(),
 verbose = FALSE, api_url = NULL, geocodio_v = 1.6, limit = 1, ...) {
-  # https://www.geocod.io/docs/#batch-geocoding
   
   # limit the dataframe to legitimate arguments
-  address_df <- address_pack$unique[names(address_pack$unique) %in% c('address', 'street', 'city', 'state', 'postalcode')]
+  address_df <- unique_addresses[names(unique_addresses) %in% get_generic_parameters('geocodio', address_only = TRUE)]
   
   ## If single line addresses are passed then we will package them as a single list
   if ('address' %in% names(address_df)) {